+++ CHANGELOG.md
@@ -13,6 +13,8 @@ Changelog
  * added config key "bootstrap-file" which is loaded before loading plugins
  * added repository type "path-mapping"
  * added config key "repository.optimize"
+ * the config key "factory.in.file" can now be set to `null` if "factory.in.class"
+   is auto-loadable
 
 * 1.0.0-beta5 (2015-05-29)